+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - IPC subcommand to retrieve output scale (`catacomb msg scale`)
+- IPC subcommand to set gesture handle height (`catacomb msg gesture-handle`)
 
 ### Changed